harmonization_exps/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.10.0=pypi_0
astunparse=1.6.3=pypi_0
blas=1.0=mkl
bzip2=1.0.8=h516909a_3
ca-certificates=2020.12.5=ha878542_0
cachetools=4.1.1=pypi_0
cairo=1.16.0=h3fc0475_1005
certifi=2020.6.20=pypi_0
cffi=1.14.0=py38h2e261b9_0
chardet=3.0.4=pypi_0
cloudpickle=1.6.0=py_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.10.0=py_2
cytoolz=0.11.0=py38h7b6447c_0
dask-core=2.30.0=py_0
dbus=1.13.6=he372182_0
decorator=4.4.2=py_0
expat=2.2.9=he1b5a44_2
ffmpeg=4.3.1=h167e202_0
fontconfig=2.13.1=h1056068_1002
freetype=2.10.2=h5ab3b9f_0
gast=0.3.3=pypi_0
gettext=0.19.8.1=hc5be6a0_1002
glib=2.66.1=h680cd38_0
gmp=6.2.0=he1b5a44_3
gnutls=3.6.13=h79a8f9a_0
google-auth=1.22.1=pypi_0
google-auth-oauthlib=0.4.1=pypi_0
google-pasta=0.2.0=pypi_0
graphite2=1.3.13=he1b5a44_1001
grpcio=1.32.0=pypi_0
gst-plugins-base=1.14.5=h0935bb2_2
gstreamer=1.14.5=h36ae1b5_2
h5py=2.10.0=nompi_py38hafa665b_105
harfbuzz=2.7.2=hee91db6_0
hdf5=1.10.6=nompi_h3c11f04_101
icu=67.1=he1b5a44_0
idna=2.10=pypi_0
imageio=2.9.0=py_0
intel-openmp=2020.2=254
jasper=1.900.1=h07fcdf6_1006
joblib=0.16.0=py_0
jpeg=9d=h516909a_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.2.0=py38hbf85e49_0
kornia=0.3.0=pyh9f0ad1d_0
krb5=1.17.1=hfafb76e_3
lame=3.100=h14c3975_1001
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.35=h769bd43_9
libblas=3.8.0=16_mkl
libcblas=3.8.0=16_mkl
libclang=10.0.1=default_hde54327_1
libedit=3.1.20191231=h14c3975_1
libevent=2.1.10=hcdb4288_2
libffi=3.2.1=he1b5a44_1007
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libiconv=1.16=h516909a_0
liblapack=3.8.0=16_mkl
liblapacke=3.8.0=16_mkl
libllvm10=10.0.1=he513fc3_3
libopencv=4.4.0=py38_2
libpng=1.6.37=hbc83047_0
libpq=12.3=h5513abc_0
libprotobuf=3.13.0=h8b12597_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=2.32.1=h14c3975_1000
libwebp-base=1.1.0=h516909a_3
libxcb=1.13=h14c3975_1002
libxkbcommon=0.10.0=he1b5a44_0
libxml2=2.9.10=h68273f3_2
lz4-c=1.9.2=he6710b0_1
markdown=3.3=pypi_0
matplotlib=3.3.2=0
matplotlib-base=3.3.2=py38h91b0d89_0
mkl=2020.2=256
mkl-service=2.3.0=py38he904b0f_0
mkl_fft=1.1.0=py38h23d657b_0
mkl_random=1.1.1=py38h0573a6f_0
multicoretsne=0.1=pyh1d8a796_0
mysql-common=8.0.21=2
mysql-libs=8.0.21=hf3661c5_2
ncurses=6.2=he6710b0_1
nettle=3.4.1=h1bed415_1002
networkx=2.5=py_0
nibabel=3.2.1=pyhd8ed1ab_0
ninja=1.10.1=py38hfd86e86_0
nspr=4.29=he1b5a44_0
nss=3.57=he751ad9_0
numpy=1.19.2=pypi_0
numpy-base=1.19.1=py38hfa32c7d_0
oauthlib=3.1.0=pypi_0
olefile=0.46=py_0
opencv=4.4.0=py38_2
openh264=2.1.1=h8b12597_0
openssl=1.1.1k=h27cfd23_0
opt-einsum=3.3.0=pypi_0
packaging=20.9=pyh44b312d_0
pandas=1.2.3=py38ha9443f7_0
pcre=8.44=he1b5a44_0
pillow=7.2.0=py38hb39fc2d_0
pip=20.2.2=py38_0
pixman=0.38.0=h516909a_1003
protobuf=3.13.0=pypi_0
pthread-stubs=0.4=h14c3975_1001
py-opencv=4.4.0=py38h23f93f0_2
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.20=py_2
pydicom=2.1.2=pyhd3deb0d_0
pyparsing=2.4.7=pyh9f0ad1d_0
python=3.8.6=h852b56e_0_cpython
python-dateutil=2.8.1=py_0
python_abi=3.8=1_cp38
pytorch=1.6.0=py3.8_cuda10.2.89_cudnn7.6.5_0
pytz=2020.4=pyhd3eb1b0_0
pywavelets=1.1.1=py38h7b6447c_2
pyyaml=5.3.1=py38h7b6447c_1
qt=5.12.9=h1f2b2cb_0
readline=8.0=h7b6447c_0
requests=2.24.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.6=pypi_0
scikit-image=0.17.2=py38hdf5156a_0
scikit-learn=0.23.2=py38h0573a6f_0
scipy=1.4.1=pypi_0
seaborn=0.11.0=py_0
setuptools=50.3.0=pypi_0
six=1.15.0=py_0
sqlite=3.33.0=h62c20be_0
tensorboard=2.2.2=pypi_0
tensorboard-plugin-wit=1.7.0=pypi_0
tensorboardx=2.1=py_0
tensorflow=2.2.0=pypi_0
tensorflow-estimator=2.2.0=pypi_0
termcolor=1.1.0=pypi_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tifffile=2020.10.1=py38hdd07704_2
tk=8.6.10=hbc83047_0
toolz=0.11.1=py_0
torchvision=0.7.0=py38_cu102
tornado=6.0.4=py38h1e0a361_1
tqdm=4.50.1=pyh9f0ad1d_0
urllib3=1.25.10=pypi_0
werkzeug=1.0.1=pypi_0
wheel=0.35.1=pypi_0
wrapt=1.12.1=pypi_0
x264=1!152.20180806=h14c3975_0
xorg-kbproto=1.0.7=h14c3975_1002
xorg-libice=1.0.10=h516909a_0
xorg-libsm=1.2.3=h84519dc_1000
xorg-libx11=1.6.12=h516909a_0
xorg-libxau=1.0.9=h14c3975_0
xorg-libxdmcp=1.1.3=h516909a_0
xorg-libxext=1.3.4=h516909a_0
xorg-libxrender=0.9.10=h516909a_1002
xorg-renderproto=0.11.1=h14c3975_1002
xorg-xextproto=7.3.0=h14c3975_1002
xorg-xproto=7.0.31=h14c3975_1007
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
